# Conflicts:
#	src/changes/changes.xml
#	src/test/java/org/orekit/files/ccsds/ndm/odm/omm/OmmParserTest.java

+++ src/test/java/org/orekit/time/UTCTAIHistoryFilesLoaderCompressedDataTest.java
@@ -1,4 +1,4 @@
-/* Copyright 2002-2024 CS GROUP
+/* Copyright 2002-2025 CS GROUP
  * Licensed to CS GROUP (CS) under one or more
  * contributor license agreements.  See the NOTICE file distributed with
  * this work for additional information regarding copyright ownership.
@@ -25,7 +25,7 @@ public class UTCTAIHistoryFilesLoaderCompressedDataTest {
     @Test
     public void testCompressed() {
         Utils.setDataRoot("compressed-data");
-        Assertions.assertEquals(-32.0, TimeScalesFactory.getUTC().offsetFromTAI(AbsoluteDate.J2000_EPOCH), 1.0e-8);
+        Assertions.assertEquals(-32.0, TimeScalesFactory.getUTC().offsetFromTAI(AbsoluteDate.J2000_EPOCH).toDouble(), 1.0e-8);
     }
 
 }
